agent-python/requirements.txt
python-socketio[client]==5.11.0
docker==7.0.0
psutil==5.9.8
